requirements.txt
flask==3.1.0
werkzeug==3.1.0
pypdf==4.0.1
google-genai==1.0.0
earthengine-api==1.0.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
numpy==2.1.3
scipy==1.14.1
matplotlib==3.9.3
pillow==11.0.0
requests==2.32.3